refactor: change song types and create song response
