+++ src/model/song.ts
@@ -1,0 +1,12 @@
+type Thumbnail = { url: string; alt: string };
+type Audio = { url: string; seconds: number };
+
+export type Song = {
+  id: number;
+  title: string;
+  description: string;
+  genre: string;
+  image: Thumbnail;
+  author: string;
+  audio: Audio;
+};
