--- src/types/audio.types.ts
@@ -1,4 +1,0 @@
-export type Audio = {
-  id: number;
-  url: string;
-};
