--- src/types/author.types.ts
@@ -1,4 +1,0 @@
-export type Author = {
-  id: number;
-  name: string;
-};
